+++ tests/dummy/app/controllers/components-examples/flexberry-lookup/settings-example.js
@@ -94,6 +94,16 @@ export default EditFormController.extend({
   */
   chooseButtonClass: '',
 
+  
+  /**
+    Text for 'flexberry-lookup' component 'dropdownClass' property.
+
+    @property chooseButtonClass
+    @type String
+    @default 'purple'
+  */
+  dropdownClass: 'negative',
+
   /**
     Text for 'flexberry-lookup' component 'removeButtonClass' property.
 
@@ -200,6 +210,13 @@ export default EditFormController.extend({
       settingDefaultValue: '',
       settingAvailableItems: ['purple basic', 'negative', 'yellow colored'],
       bindedControllerPropertieName: 'removeButtonClass'
+    });
+    componentSettingsMetadata.pushObject({
+      settingName: 'dropdownClass',
+      settingType: 'css',
+      settingDefaultValue: '',
+      settingAvailableItems: ['blue'],
+      bindedControllerPropertieName: 'dropdownClass'
     });
 
     return componentSettingsMetadata;